remove redundant toLong

+++ core/src/main/scala/org/apache/spark/SparkContext.scala
@@ -720,7 +720,7 @@ class SparkContext(config: SparkConf) extends Logging with ExecutorAllocationCli
       }
     }
     parallelize(0 until numSlices, numSlices).mapPartitionsWithIndex((i, _) => {
-      val partitionStart = ((i * numElements) / numSlices * step + start).toLong
+      val partitionStart = (i * numElements) / numSlices * step + start
       val partitionEnd = (((i + 1) * numElements) / numSlices) * step + start
       def getSafeMargin(bi: BigInt): Long =
         if (bi.isValidLong) {
